+++ R/class_queue.R
@@ -30,6 +30,9 @@ queue_class <- R6::R6Class(
     },
     dequeue = function() {
     },
+    abridge = function() {
+      while (self$is_nonempty()) self$dequeue()
+    },
     prepend = function(names, ranks = NULL) {
     },
     increment_ranks = function(names, by) {