+++ R/tar_target.R
@@ -137,11 +137,11 @@
 #'   way to save and load this file. See the "Storage formats" section
 #'   for a detailed list of possible data storage formats.
 #' @param error Character of length 1, what to do if the target
-#'   runs into an error. If `"stop"`, the whole pipeline stops
-#'   and throws an error. If `"continue"`, the error is recorded,
-#'   but the pipeline keeps going. `error = "workspace"` is just like
-#'   `error = "stop"` except `targets` saves a special workspace file
-#'   to support interactive debugging outside the pipeline.
+#'   stops and throws an error. Options:
+#'   * `"stop"`: the whole pipeline stops and throws an error.
+#'   * `"continue"`: the whole pipeline keeps going.
+#'   * `"abridge"`: any currently running targets keep running,
+#'     but no new targets launch after that.
 #'   (Visit <https://books.ropensci.org/targets/debugging.html>
 #'   to learn how to debug targets using saved workspaces.)
 #' @param memory Character of length 1, memory strategy.