+++ tests/testthat/test-tar_workspace.R
@@ -10,7 +10,7 @@ tar_test("workspaces are not saved if error = 'stop'", {
   expect_false(file.exists(path_workspace(path_store_default(), "x")))
 })
 
-tar_test("workspaces are not saved if error = 'continue'", {
+tar_test("workspaces are not saved by default", {
   pipeline <- pipeline_init(
     list(
       target_init("y", quote(12345)),
@@ -23,7 +23,20 @@ tar_test("workspaces are not saved if error = 'continue'", {
   expect_false(file.exists(path_workspace(path_store_default(), "x")))
 })
 
-tar_test("workspaces are saved if error = 'save'", {
+tar_test("workspaces are saved if requested on error", {
+  tar_option_set(workspaces = tar_workspace_policy(error = TRUE))
+  pipeline <- pipeline_init(
+    list(
+      target_init("y", quote(1)),
+      target_init("x", quote(stop(y)))
+    )
+  )
+  local <- local_init(pipeline, reporter = "verbose")
+  expect_error(expect_message(local$run()), class = "tar_condition_run")
+  expect_true(file.exists(path_workspace(path_store_default(), "x")))
+})
+
+tar_test("deprecated workspace error option works", {
   pipeline <- pipeline_init(
     list(
       target_init("y", quote(1)),
@@ -35,10 +48,41 @@ tar_test("workspaces are saved if error = 'save'", {
   expect_true(file.exists(path_workspace(path_store_default(), "x")))
 })
 
+tar_test("deprecated workspaces option works", {
+  suppressWarnings(tar_option_set(workspaces = "x"))
+  pipeline <- pipeline_init(
+    list(
+      target_init("y", quote(1)),
+      target_init("x", quote(y))
+    )
+  )
+  local <- local_init(pipeline, reporter = "verbose")
+  local$run()
+  expect_true(file.exists(path_workspace(path_store_default(), "x")))
+})
+
+tar_test("never arg overrides error arg", {
+  tar_option_set(
+    workspaces = tar_workspace_policy(
+      never = "x",
+      error = TRUE
+    )
+  )
+  pipeline <- pipeline_init(
+    list(
+      target_init("y", quote(1)),
+      target_init("x", quote(stop(y)))
+    )
+  )
+  local <- local_init(pipeline, reporter = "verbose")
+  expect_error(expect_message(local$run()), class = "tar_condition_run")
+  expect_false(file.exists(path_workspace(path_store_default(), "x")))
+})
+
 tar_test("tar_workspace() works", {
   tmp <- sample(1)
   tar_script({
-    tar_option_set(error = "workspace")
+    tar_option_set(workspaces = tar_workspace_policy(error = TRUE))
     list(
       tar_target(x, "loaded"),
       tar_target(y, stop(x))
@@ -55,7 +99,7 @@ tar_test("tar_workspace() works", {
 tar_test("tar_workspace() works with workspace option", {
   tmp <- sample(1)
   tar_script({
-    tar_option_set(workspaces = "y")
+    tar_option_set(workspaces = tar_workspace_policy(always = "y"))
     list(
       tar_target(x, "loaded"),
       tar_target(y, paste0(x, "nope"))
@@ -69,7 +113,7 @@ tar_test("tar_workspace() works with workspace option", {
 
 tar_test("tar_workspace() on a branch", {
   tar_script({
-    tar_option_set(error = "workspace")
+    tar_option_set(workspaces = tar_workspace_policy(error = TRUE))
     list(
       tar_target(x, seq_len(4L)),
       tar_target(y, stopifnot(x < 4L), pattern = map(x))
@@ -86,11 +130,25 @@ tar_test("tar_workspace() on a branch", {
   expect_equal(envir$x, 4L)
 })
 
+tar_test("tar_workspace() on a pattern", {
+  tar_script({
+    tar_option_set(workspaces = tar_workspace_policy(always = "y"))
+    list(
+      tar_target(x, seq_len(2L)),
+      tar_target(y, x, pattern = map(x))
+    )
+  })
+  try(tar_make(callr_function = NULL), silent = TRUE)
+  branches <- tar_branch_names("y", seq_len(2))
+  result <- file.exists(path_workspace(path_store_default(), branches))
+  expect_true(all(result))
+})
+
 tar_test("tar_workspace() with an unexportable object", {
   skip_on_cran()
   skip_if_not_installed("torch")
   tar_script({
-    tar_option_set(error = "workspace")
+    tar_option_set(workspaces = tar_workspace_policy(error = TRUE))
     list(
       tar_target(tensor, torch::torch_zeros(10), format = "torch"),
       tar_target(array, stop(tensor))
@@ -111,7 +169,7 @@ tar_test("workspace saved on no error and when target is skipped", {
   tar_make(callr_function = NULL)
   expect_false(file.exists(path))
   tar_script({
-    tar_option_set(workspaces = "z")
+    tar_option_set(workspaces = tar_workspace_policy(always = "z"))
     list(tar_target(z, 0))
   })
   tar_make(callr_function = NULL)
@@ -123,7 +181,7 @@ tar_test("custom script and store args", {
   expect_equal(tar_config_get("script"), path_script_default())
   expect_equal(tar_config_get("store"), path_store_default())
   tar_script({
-    tar_option_set(error = "workspace")
+    tar_option_set(workspaces = tar_workspace_policy(error = TRUE))
     list(tar_target(x, "value"), tar_target(y, stop(x)))
   }, script = "example/script.R")
   try(